Fix duplicate Q_NAMESPACE issue
Having two instances of Q_NAMESPACE causes issues with Qt, so move
all enums into one file.
