+++ sealtk/gui/Enums.cpp
@@ -1,0 +1,1 @@
+// This file exists solely to trigger automoc for the corresponding header
